Revise generation of functions related to Xcoding error
Previously, the error_decode was generated only when at least one
declared error contains an extra information.

As the standard considers this extra information as a nullable Element,
the error_decode function is now always generated. If the extra
information is not provided, the returned mal_element_holder will
have its presence flag set to false.

Furthermore, the encoding functions related to error are generated
only for the types of the declared extra informations.